# Fix FP4 MoE accuracy from missing routed_scaling_factor