@fzyzcjy 2025-07-25
I also see accuracy drop for fp4 like <90% acc. wondering whether this will be useful!