@fzyzcjy 2025-07-26
Hi, could you please share the lmeval command you use to test gsm8k, then I can be aligned with this